+++ examples/08_thermocouple/hal_project_config.h
@@ -1,0 +1,4 @@
+#pragma once
+
+#define HAL_ENABLE_MCP9600
+#define HAL_ENABLE_MAX6675