+++ examples/09_display_tft/hal_project_config.h
@@ -1,0 +1,4 @@
+#pragma once
+
+#define HAL_ENABLE_ILI9341
+#define HAL_DISPLAY_ILI9341