+++ examples/12_kv_store/hal_project_config.h
@@ -1,0 +1,4 @@
+#pragma once
+
+#define HAL_ENABLE_KV
+#define HAL_ENABLE_I2C