+++ examples/13_i2c_slave/hal_project_config.h
@@ -1,0 +1,3 @@
+#pragma once
+
+#define HAL_ENABLE_I2C_SLAVE