+++ examples/16_littlefs/hal_project_config.h
@@ -1,0 +1,3 @@
+#pragma once
+
+#define HAL_ENABLE_LITTLEFS