+++ examples/18_rgb_led/hal_project_config.h
@@ -1,0 +1,3 @@
+#pragma once
+
+#define HAL_ENABLE_RGB_LED